refactor: move the non-hook usePrice into util
