--- app/utils/price-formtter.ts
@@ -1,8 +1,0 @@
-export const formatCurrency = (
-  price: number | string = 0,
-  currency: string = 'USD'
-) =>
-  Intl.NumberFormat('en-US', {
-    style: 'currency',
-    currency,
-  }).format(Number(price));
